cuspatial.h renamed to cuspatial/types.hpp
