+++ cpp/include/cuspatial/soa_readers.hpp
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <cudf/cudf.h>
-#include <cuspatial/cuspatial.h>
+#include <cuspatial/types.hpp>
 
 namespace cuspatial {
