+++ cpp/include/cuspatial/types.hpp
