+++ cpp/src/include/trajectory_thrust.cuh
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <ostream>
-#include <cuspatial/cuspatial.h>
+#include <cuspatial/types.hpp>
 
 namespace cuspatial
 {
